ai책추천 파일 수정 및 마이페이지 기능 추가

+++ src/main/java/com/bookreum/dev/domain/book/AladinItem.java
@@ -7,11 +7,11 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AladinItem {
-    private Long itemId;
-    private String title;
-    private String author;
-    private String cover;
-    private String isbn13;
-    private String categoryName;
-    private String description;
+	private Integer   itemId;        // 알라딘 내부 식별자
+    private String title;         // 도서 제목
+    private String author;        // 저자명
+    private String cover;         // 표지 이미지 URL
+    private String isbn13;        // ISBN-13
+    private String categoryName;  // 카테고리 이름
+    private String description;   // 간략 소개
 }
